bsps/or1k: Address -Wsign-compare warnings
This warning occurs when comparing a signed variable to an unsigned one.
This addresses warnings that only occurred on 64-bit targets.  For the
ones which only appeared on 64-bit targets, the cause was frequently
a mismatch when comparing a combination off_t, ssize_t, and int.

+++ bsps/or1k/shared/cache/cache.c
@@ -207,7 +207,7 @@ static void _CPU_cache_flush_data_range(const void *d_addr, size_t n_bytes)
   final_address = (void *)((size_t)d_addr + n_bytes - 1);
   d_addr = (void *)((size_t)d_addr & ~(CPU_DATA_CACHE_ALIGNMENT - 1));
 
-  if( final_address - d_addr > _CPU_cache_get_data_cache_size(0) ) {
+  if( (size_t)(final_address - d_addr) > _CPU_cache_get_data_cache_size(0) ) {
     /*
      * Avoid iterating over the whole cache multiple times if the range is
      * larger than the cache size.
@@ -244,7 +244,7 @@ static void _CPU_cache_invalidate_data_range(const void *d_addr, size_t n_bytes)
   final_address = (void *)((size_t)d_addr + n_bytes - 1);
   d_addr = (void *)((size_t)d_addr & ~(CPU_DATA_CACHE_ALIGNMENT - 1));
 
-  if( final_address - d_addr > _CPU_cache_get_data_cache_size(0) ) {
+  if( (size_t)(final_address - d_addr) > _CPU_cache_get_data_cache_size(0) ) {
     /*
      * Avoid iterating over the whole cache multiple times if the range is
      * larger than the cache size.
@@ -281,7 +281,7 @@ static void _CPU_cache_invalidate_instruction_range(const void *i_addr, size_t n
   final_address = (void *)((size_t)i_addr + n_bytes - 1);
   i_addr = (void *)((size_t)i_addr & ~(CPU_INSTRUCTION_CACHE_ALIGNMENT - 1));
 
-  if( final_address - i_addr > _CPU_cache_get_data_cache_size(0) ) {
+  if( (size_t)(final_address - i_addr) > _CPU_cache_get_data_cache_size(0) ) {
     /*
      * Avoid iterating over the whole cache multiple times if the range is
      * larger than the cache size.
